refactor: réorganisation de l'arborescence back-end (#469)
Comme convenu, voici un PR de refactoring qui effectue les changements
suivants :

- retravailler l'arborescence pour avoir [modéle] / [couches d'archi
DDD] plutôt que l'inverse
- Renommer "helpers" en "common"
- Renommer les fichiers en camel case
- Renommer "domain/services" en "application"
- Découper par anticipation les dossier "infrastructure" en fonction de
leur nature ("api/brevo", "api/sirene")

Au passage 2 petits changements orthogonaux : 
- configurer jest pour ne lancer que les tests du répertoire "tests"
dans packages/backend
- remplacer `jest.toThrowError` (déprécié) par `jest.toThrow` 

J'ai corrigé les imports en conséquence.
