+++ packages/backend/src/common/jsonError.ts
