+++ packages/backend/src/contact/infrastructure/api/brevo/types.ts
@@ -1,0 +1,5 @@
+export interface BrevoBody {
+  email: string
+  listIds: number[]
+  attributes: object
+}
