--- packages/backend/src/domain/services/program-service.ts
@@ -1,6 +1,0 @@
-import { currentDateService } from '../../infrastructure/current-date'
-import { createService as createFilterService } from '../program/filter-programs'
-
-export { sortPrograms } from '../program/sort-programs'
-
-export const filterPrograms = createFilterService(currentDateService)
