+++ packages/backend/src/program/application/sortAndFilterPrograms.ts
@@ -1,0 +1,6 @@
+import { currentDateService } from '../infrastructure/currentDate'
+import { createService as createFilterService } from '../domain/filterPrograms'
+
+export { sortPrograms } from '../domain/sortPrograms'
+
+export const filterPrograms = createFilterService(currentDateService)
